==> ICU/01-Download_data.ipynb <==
# Installation librairies

# Import librairies

import pandas as pd
import requests
import os

# DATA IMPORT:

from google.colab import drive
drive.mount('/content/drive')

# chemins
path = '/content/drive/MyDrive/ICU/'
path_down = '/content/drive/MyDrive/ICU/Download_data/'
path_analyse = '/content/drive/MyDrive/ICU/Analyse/'
path_carto = '/content/drive/MyDrive/ICU/Carto/'

# Importer depuis une URL API le csv des stations en place vers Google drive

# Configuration
API_URL = 'https://data.toulouse-metropole.fr/api/explore/v2.1/catalog/datasets/stations-meteo-en-place/exports/csv?lang=fr&timezone=Europe%2FBerlin&use_labels=true&delimiter=%3B'

# Chemin complet du fichier où sauvegarder le CSV
file_path = path_down+'Liste_station.csv'

# Télécharger les données CSV
def download_csv(api_url, file_path):
    response = requests.get(api_url)
    response.raise_for_status()  # Assure que la requête a réussi
    with open(file_path, 'wb') as file:
        file.write(response.content)
    print(f"CSV file downloaded and saved as {file_path}")

# Exécuter les fonctions
if __name__ == '__main__':
    download_csv(API_URL, file_path)

# Import Dataset

# import liste station csv
df = pd.read_csv(path_down+'Liste_station.csv',sep=';')
# creation liste des stations valides (emission = V)
Liste_stations = sorted(df[df['emission']!='N']['id_nom'].unique().tolist())
# transformation en minuscules
Liste_stations = [station.lower() for station in Liste_stations]
len(Liste_stations)

# URL de base de l'API
base_api_url_part1 = 'https://data.toulouse-metropole.fr/api/explore/v2.1/catalog/datasets/'
base_api_url_part2 = '/exports/csv?lang=fr&timezone=Europe%2FBerlin&use_labels=true&delimiter=%3B'

# Fonction pour télécharger le CSV avec gestion d'erreurs
def download_csv_e(api_url, file_path):
    try:
        response = requests.get(api_url)
        response.raise_for_status()  # Vérifie si la requête est réussie
        with open(file_path, 'wb') as file:
            file.write(response.content)
        print(f"Fichier CSV téléchargé et sauvegardé sous {file_path}")
    except requests.exceptions.HTTPError as http_err:
        if response.status_code == 404:
            print(f"Erreur 404: La station n'existe pas à l'URL : {api_url}")
        else:
            print(f"Erreur HTTP: {http_err}")
    except Exception as err:
        print(f"Une erreur s'est produite : {err}")

# Boucle sur les valeurs distinctes pour télécharger les fichiers
for station in Liste_stations:
    # Adapter l'URL et le nom de fichier en fonction de la station
    api_url = f"{base_api_url_part1}{station}{base_api_url_part2}"
    file_path = os.path.join(path_down, f'{station}.csv')

    # Télécharger les données CSV
    download_csv_e(api_url, file_path)
