==> ICU/02-Cleaner_data.ipynb <==
# Installation librairies

#

# Import librairies

import pandas as pd
import requests
import os

# DATA IMPORT:

from google.colab import drive
drive.mount('/content/drive')

# chemins
path = '/content/drive/MyDrive/ICU/'
path_down = '/content/drive/MyDrive/ICU/Download_data/'
path_analyse = '/content/drive/MyDrive/ICU/Analyse/'
path_carto = '/content/drive/MyDrive/ICU/Carto/'

# Nettoyage dataset

# Lister les datasets
List_dataset = [f for f in os.listdir(path_down) if os.path.isfile(os.path.join(path_down, f))]

# Fichiers à supprimer de la liste
files_to_remove = ['Liste_station.csv']

# Utilisation d'une boucle pour supprimer les fichiers de la liste s'ils existent
for file in files_to_remove:
    if file in List_dataset:
        List_dataset.remove(file)

# Supprimer le '.csv' dans les noms de fichiers
List_dataset = [f.replace('.csv', '') for f in List_dataset]

print(List_dataset)

## Import Dataset

# Initialiser un dictionnaire pour stocker les DataFrames
dfs = {}

# Liste des colonnes à conserver et organiser
colonnes_a_conserver = [
    'heure_utc', 'id', 'type_de_station', 'temperature', 'humidite', 'pression',
    'pluie', 'pluie_intensite_max', 'direction_du_vecteur_vent_moyen',
    'force_moyenne_du_vecteur_vent', 'direction_du_vecteur_de_vent_max',
    'direction_du_vecteur_de_rafale_de_vent_max', 'force_rafale_max'
]

# Dictionnaire pour spécifier les types de données des colonnes
dtype_dict = {
    'heure_utc': 'str',
    'station_name': 'str',
    'id': 'str',
    'type_de_station': 'str',
    'temperature': 'float64',
    'humidite': 'float64',
    'pression': 'float64',
    'pluie': 'float64',
    'pluie_intensite_max': 'float64',
    'direction_du_vecteur_vent_moyen': 'float64',
    'force_moyenne_du_vecteur_vent': 'float64',
    'direction_du_vecteur_de_vent_max': 'float64',
    'direction_du_vecteur_de_rafale_de_vent_max': 'float64',
    'force_rafale_max': 'float64',
}

# Boucle pour créer un DataFrame pour chaque fichier dans List_dataset
for dataset in List_dataset:
    file_path = os.path.join(path_down, dataset+'.csv')
    try:
        # Lire chaque fichier CSV et stocker dans le dictionnaire avec les types de données spécifiés
        dfs[f"df_{dataset}"] = pd.read_csv(file_path, sep=';', dtype=dtype_dict, low_memory=False)

        # Vérifier et renommer les colonnes si nécessaire
        if 'dd_mm_yy_hh_mm' in dfs[f"df_{dataset}"].columns:
            dfs[f"df_{dataset}"].rename(columns={'dd_mm_yy_hh_mm': 'heure_utc'}, inplace=True)
        if 'temperature_en_degre_c' in dfs[f"df_{dataset}"].columns:
            dfs[f"df_{dataset}"].rename(columns={'temperature_en_degre_c': 'temperature'}, inplace=True)

        # Vérifier les colonnes présentes et organiser selon celles disponibles
        colonnes_disponibles = [col for col in colonnes_a_conserver if col in dfs[f"df_{dataset}"].columns]
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"][colonnes_disponibles]

        # Supprimer les doublons dans les colonnes
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"].loc[:, ~dfs[f"df_{dataset}"].columns.duplicated()]

        # Supprimer les lignes avec des valeurs manquantes
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"].dropna()

        # Supprimer les doublons
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"].drop_duplicates()

        # Identifier les valeurs aberrantes pour la colonne 'temperature'
        q1 = dfs[f"df_{dataset}"]['temperature'].quantile(0.25)
        q3 = dfs[f"df_{dataset}"]['temperature'].quantile(0.75)
        iqr = q3 - q1
        lower_bound = q1 - 1.5 * iqr
        upper_bound = q3 + 1.5 * iqr

        # Filtrer les valeurs aberrantes
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"][(dfs[f"df_{dataset}"]['temperature'] >= lower_bound) & (dfs[f"df_{dataset}"]['temperature'] <= upper_bound)]

        # Convertir la colonne 'heure_utc' en type datetime
        if 'heure_utc' in dfs[f"df_{dataset}"].columns:
            dfs[f"df_{dataset}"].loc[:, 'heure_utc'] = pd.to_datetime(dfs[f"df_{dataset}"].loc[:, 'heure_utc'], utc=True)

        # Trier par 'heure_utc'
        if 'heure_utc' in dfs[f"df_{dataset}"].columns:
            dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"].sort_values(by='heure_utc')

        # Droper les lignes où 'id' est différent de la valeur la plus fréquente
        if 'id' in dfs[f"df_{dataset}"].columns:
            id_plus_frequent = dfs[f"df_{dataset}"]['id'].mode()[0]  # Trouver l'ID le plus fréquent
            dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"][dfs[f"df_{dataset}"]['id'] == id_plus_frequent]  # Filtrer selon cet ID

        # Ajouter le nom de la station
        dfs[f"df_{dataset}"]['station_name'] = f"{dataset}"
        # Placer la colonne station_name en premier
        cols = ['station_name'] + [col for col in dfs[f"df_{dataset}"].columns if col != 'station_name']
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"][cols]

        # Réinitialiser les index après toutes les opérations
        dfs[f"df_{dataset}"] = dfs[f"df_{dataset}"].reset_index(drop=True)

        print(f"DataFrame créé, organisé et trié pour {dataset}")

    except Exception as e:
        print(f"Erreur sur {dataset} : {e}")

# Concaténer tous les DataFrames dans df_Stations_All
df_Stations_All = pd.concat(dfs.values(), ignore_index=True)

# Exporter df_Stations_All en un fichier CSV
output_path = os.path.join(path_analyse, 'Stations_All.csv')  # Spécifier le chemin de sauvegarde du fichier CSV
df_Stations_All.to_csv(output_path, sep=';', index=False)

print(f"DataFrame exporté en CSV à : {output_path}")

display(df_Stations_All.head(5))
df_Stations_All.describe()
