==> ICU/03-ICU_Station_select.ipynb <==
# Installation librairies

# Import librairies

import pandas as pd
import os

# DATA IMPORT:

from google.colab import drive
drive.mount('/content/drive')

# chemins
path = '/content/drive/MyDrive/ICU/'
path_down = '/content/drive/MyDrive/ICU/Download_data/'
path_analyse = '/content/drive/MyDrive/ICU/Analyse/'
path_carto = '/content/drive/MyDrive/ICU/Carto/'

# Import Dataset

# import Stations data csv
df = pd.read_csv(path_analyse+'Stations_All.csv',sep=';', low_memory=False)

display(df.info())
display(df.head(2))

# Transformation & select 2023 data

# transforme la colonne 'date' au format datetime
df['heure_utc'] = pd.to_datetime(df['heure_utc'], utc=True)

# format la colonne 'id' en string
df['id'] = df['id'].astype(str)

# Filtrer les lignes où l'année est = à 2023
df_2023 = df[df['heure_utc'].dt.year == 2023].sort_values(by=['id','heure_utc'])

# Filtrer les valeurs de températures aberrante

df_2023 = df_2023[df_2023['temperature']> -20]

df_2023.info()

# Resample sur un intervalle de 3 heures

# Fixer 'heure_utc' comme index
df_2023.set_index('heure_utc', inplace=True)

# Séparer les colonnes numériques et non numériques (sans 'id' pour les non-numériques)
cols_num = df_2023.select_dtypes(include=['number']).columns
cols_non_num = df_2023.select_dtypes(exclude=['number']).columns.difference(['id'])

# Resample uniquement les colonnes numériques sur un intervalle de 3 heures
df_numeric_resampled = df_2023[cols_num].groupby(df_2023['id']).resample('3h').mean(numeric_only=True)

# Récupérer les colonnes non numériques (comme id) sans affecter le resampling
df_non_numeric = df_2023[cols_non_num].groupby(df_2023['id']).resample('3h').first()

# Fusionner les résultats numériques et non numériques
df_2023_3h = pd.concat([df_numeric_resampled, df_non_numeric], axis=1).reset_index()

df_2023_3h.head()

# Groupby sur 'id' puis compter les occurrences des valeurs de 'temperature'
temperature_counts = df_2023_3h.groupby('id')['temperature'].count().sort_values(ascending=False)

#seuil 2628
seuil = 2569

# Filtre les stations ayant de résultats (8 mesures par jour sur 365 jours x 0.88 = 2570)
stations_to_keep = temperature_counts[temperature_counts >= seuil ].index

df_stations_to_keep = df_2023_3h[df_2023_3h['id'].isin(stations_to_keep)]

display(temperature_counts)

#list des id de stations + count
df_stations_to_keep['id'].unique(), df_stations_to_keep['id'].nunique()

# Calcul du delta de T° avec la station ref

# 1. Extraire les lignes de la station référence id = 31
df_ref = df_stations_to_keep[df_stations_to_keep['id'] == '31']

# 2. Fusionner les lignes de référence avec le reste sur la base de la colonne 'timestamp'
df_merged = df_stations_to_keep.merge(df_ref[['heure_utc', 'temperature']], on='heure_utc', suffixes=('', '_ref'))

# 3. Calculer le delta de température
df_merged['delta_temperature'] = df_merged['temperature'] - df_merged['temperature_ref']

# 4. Filtrer pour ne conserver que les lignes où id != 1 (puisque id = 1 serait delta = 0)
df_delta = df_merged[df_merged['id'] != 1]

# Afficher le résultat
display(df_delta[['id', 'heure_utc', 'temperature', 'temperature_ref', 'delta_temperature']].head(5))

# Compter le nombre de fois où delta_temperature est supérieur à 4 pour chaque id
delta_count = df_delta[df_delta['delta_temperature'] >= 4].groupby('id').size().reset_index(name='compteur')
#display(delta_count)

# Fusionner ce comptage avec le DataFrame d'origine
df_with_compteur = df_delta.merge(delta_count, on='id', how='left')

# Compter le nombre total d'occurrences pour chaque id dans le DataFrame d'origine
total_count = df_delta[~df_delta['temperature'].isnull()].groupby('id').size().reset_index(name='total')

# Remplir les NaN dans la colonne 'compteur' avec 0 pour les ids qui n'ont aucun delta supérieur à 4
df_with_compteur['compteur'] = df_with_compteur['compteur'].fillna(0).astype(int)

# Fusionner ce comptage avec le DataFrame d'origine
df_with_compteur_total = df_with_compteur.merge(total_count, on='id', how='left')

#Ratio
df_with_compteur_total['Ratio'] = df_with_compteur_total['compteur']/df_with_compteur_total['total']

# Afficher le DataFrame mis à jour
display(df_with_compteur_total.head(2))

# Sélectionner uniquement les colonnes numériques pour calculer la moyenne
Top_3 = df_with_compteur_total.groupby('id').mean(numeric_only=True).sort_values('Ratio', ascending=False).head(4)
Less_3 = df_with_compteur_total.groupby('id').mean(numeric_only=True).sort_values('Ratio', ascending=False).tail(4)

# Afficher les résultats
display(Top_3)
display(Less_3)

# Exporter dataset_ICU en un fichier CSV
output_path = os.path.join(path_analyse, 'dataset_ICU.csv')  # Spécifier le chemin de sauvegarde du fichier CSV
df_with_compteur_total.to_csv(output_path, sep=';', index=False)

print(f"DataFrame exporté en CSV à : {output_path}")
